requirements.txt
# Core dependencies
apache-beam[gcp]==2.54.0
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
google-cloud-functions==1.16.0
pydantic==2.5.1
pydantic[email]==2.5.1

# Claude LLM integration
anthropic==0.7.8
requests==2.31.0

# Data processing
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
google-cloud-logging==3.8.0
structlog==23.2.0

# Utilities
click==8.1.7
jsonschema==4.20.0 